refactor verify code out of command file
